MK-1359 Base study class added to prepare harmonization study
